Create math_facts_v1.py
This is a very flexible program. The questions vary, and you will be told your score, as well as if you got a question right. It also tells you what question you are on, in case you are doing many math questions. However, it only supports addition because if you were to change the math operation, you'd have to change the code. If you wanted to change the number of facts, you would also need to change the code.